refactor: change location of nav comps
