+++ src/components/nav/language-switcher.tsx
@@ -13,7 +13,7 @@ import {
 	DropdownMenuSub,
 	DropdownMenuSubContent,
 	DropdownMenuSubTrigger
-} from "./ui/dropdown-menu";
+} from "../ui/dropdown-menu";
 
 export default function LanguageSwitcher() {
 	const t = useTranslations();
